requirements.txt
fastapi
uvicorn
langchain-google-genai
langgraph
requests
pydantic
python-dotenv
langchain_openai
